# Optimized ref filling.