--- a/fill.go
+++ b/fill.go
@@ -55,19 +55,13 @@ func fillRefs(impl any, get func(reflect.Type) (any, error)) error {
 
 	for i, n := 0, s.NumField(); i < n; i++ {
 		f := s.Field(i)
-		if !f.Type().Implements(reflection.Type[interface{ isRef() }]()) {
+		if !f.CanAddr() {
 			continue
 		}
-
-		// Sanity check that field type structure matches weaver.Ref[T].
-		if f.Kind() != reflect.Struct {
-			panic(fmt.Errorf("FillRefs: weaver.Ref %v is not a struct", f))
-		}
-		if f.NumField() != 1 {
-			panic(fmt.Errorf("FillRefs: weaver.Ref %v does not have one field", f))
-		}
-		if f.Type().Field(0).Name != "value" {
-			panic(fmt.Errorf("FillRefs: weaver.Ref %v field 0 not named %q", f, "value"))
+		p := reflect.NewAt(f.Type(), f.Addr().UnsafePointer()).Interface()
+		x, ok := p.(interface{ setRef(any) })
+		if !ok {
+			continue
 		}
 
 		// Set the component.
@@ -76,7 +70,7 @@ func fillRefs(impl any, get func(reflect.Type) (any, error)) error {
 		if err != nil {
 			return fmt.Errorf("FillRefs: setting field %v.%s: %w", s.Type(), s.Type().Field(i).Name, err)
 		}
-		setPossiblyUnexported(valueField, reflect.ValueOf(component))
+		x.setRef(component)
 	}
 	return nil
 }
@@ -129,13 +123,4 @@ func getConfig(impl any) any {
 		return c.getConfig()
 	}
 	return nil
-}
-
-// setPossiblyUnexported sets dst to value. It is equivalent to
-// reflect.Value.Set except that it works even if dst was accessed via
-// unexported fields.
-func setPossiblyUnexported(dst, value reflect.Value) {
-	// Use UnsafePointer + NewAt so we can handle unexported fields.
-	dst = reflect.NewAt(dst.Type(), dst.Addr().UnsafePointer()).Elem()
-	dst.Set(value)
 }